LNP now supports loading a DF installation from an arbitrary location.
This feature is very useful on Linux, as now the user doesn't have to put DF in the same place as LNP. We automatically look for DF installations in . and /opt

This was quite an invasive change. Since the DF install is no longer hard coded, every UI component must be prepared for the folder to change.

I used a singleton pattern in the new DwarfFortress class to ease and control access to the DF installation information.